refactor: type defines
