+++ wsdf/src/wireshark/mod.rs
@@ -1,3 +1,4 @@
 pub mod dissector;
 pub mod protocol;
 pub mod types;
+pub use types::*;
